Move common/turboqunat back (but keep the refactor)
