--- lib/common/turboquant/Cargo.toml
@@ -1,22 +1,0 @@
-[package]
-name = "turboquant"
-version = "0.0.0"
-authors = [
-    "Andrey Vasnetsov <andrey@vasnetsov.com>",
-    "Qdrant Team <info@qdrant.tech>",
-]
-license = "Apache-2.0"
-edition = "2024"
-publish = false
-
-[lints]
-workspace = true
-
-[dependencies]
-common = { path = "../common" }
-serde = { workspace = true }
-
-[dev-dependencies]
-rand = { workspace = true }
-rstest = { workspace = true }
-rand_distr = { workspace = true }
