+++ lib/quantization/src/lib.rs
@@ -7,6 +7,7 @@ pub mod encoded_vectors_u8;
 pub mod kmeans;
 pub mod p_square;
 pub mod quantile;
+pub mod turboquant;
 
 use std::fmt::Display;
 use std::sync::{Arc, Condvar, Mutex};
