+++ lib/quantization/src/turboquant/permutation.rs
